ALERT-51: Removed mbta api v2 and replaced it with v3.

+++ app/src/main/java/com/jesseoberstein/alert/listeners/StartActivityOnClick.java
@@ -133,7 +133,7 @@ public class StartActivityOnClick implements OnClickListener, OnItemClickListene
     }
 
     private void forwardToSelectedAlarm(CustomListItem item) {
-        UserAlarm userAlarm = ((ViewAlarms) this.origin).getUserAlarmDao().queryForId(item.getId());
+        UserAlarm userAlarm = ((ViewAlarms) this.origin).getUserAlarmDao().queryForId((int) item.getId());
         this.extras.putSerializable(ALARM, userAlarm);
         this.extras.putStringArrayList(ENDPOINTS,
                 Arrays.stream(item.getTertiaryText().split(","))
